git-svn-id: svn://192.168.1.10/public/CoffeeMud@5371 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Commands/AutoLoot.java
@@ -53,7 +53,7 @@ public class AutoLoot extends StdCommand
 		}
 		return false;
 	}
-	public int actionsCost(){return 0;}
+	public double actionsCost(){return 0.0;}
 	public boolean canBeOrdered(){return true;}
 }
