+++ com/planet_ink/coffee_mud/Commands/AutoMelee.java
@@ -57,7 +57,7 @@ public class AutoMelee extends StdCommand
 		}
 		return false;
 	}
-	public int actionsCost(){return 0;}
+	public double actionsCost(){return 0.0;}
 	public boolean canBeOrdered(){return true;}
 }
