+++ com/planet_ink/coffee_mud/Commands/AutoNotify.java
@@ -54,6 +54,6 @@ public class AutoNotify extends StdCommand
 		}
 		return false;
 	}
-	public int actionsCost(){return 0;}
+	public double actionsCost(){return 0.0;}
 	public boolean canBeOrdered(){return true;}
 }
